requirements.txt
asgiref==3.11.0
crispy-bootstrap5==2025.6
Django==6.0.1
django-crispy-forms==2.5
django-htmx==1.27.0
sqlparse==0.5.5
tzdata==2025.3
